pull key handling code out of ecdsa into a separate module
